+++ tests/test_liquidity_universe.py
@@ -1,5 +1,3 @@
-import os
-
 import pytest
 
 import pandas as pd