+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 0.6.6
+
+- New time points for the [time buckets](https://tradingstrategy.ai/docs/programming/api/timebucket.html) 
+
+# 0.6.5
+
+- Add `exchange_address` as a part of [trading pair datasets](https://tradingstrategy.ai/docs/programming/api/pair.html)
+- Add `get_by_chain_and_factory()` to [decentralised exchange universe](https://tradingstrategy.ai/docs/programming/api/exchange.html)
+
+# 0.6.4
+
+- Add `buy_tax`, `transfer_tax`, `sell_tax` to [trading pair datasets](https://tradingstrategy.ai/docs/programming/api/pair.html)
+
+# 0.6.3
+
+- Fix download retries in live trading
+- Add `stablecoin` module for upcoming stablecoin support fuctions
+- Add `DEXPair.convert_to_dataframe` 
+- Add `filter_for_stablecoins` 
+  
+
 # 0.6.2
 
 - Fix broken build