+++ docs/source/programming/release-notes.rst
@@ -1,6 +1,6 @@
 Release notes
 =============
 
-`All new releases are available on PyPi <https://pypi.org/project/tradingstrategy/>`_.
+`All new releases are available on PyPi <https://pypi.org/project/trading-strategy/>`_.
 
 See `the Change Log file on Github <https://github.com/tradingstrategy-ai/trading-strategy/blob/master/CHANGELOG.md>`_.